requirements.txt
django
djangorestframework
dotenv
requests
djangorestframework-simplejwt
python-decouple
django-cryptography
django-cors-headers
